removed emptype from the users view

+++ views/users/users.view.php
@@ -37,8 +37,7 @@ $countusers = $db->read('users');
                     <option value="<?= count($countusers) ?>" <?= ($limit == count($countusers)) ? "SELECTED" : "" ?>>
                         All</option>
                 </select>
-                <button type="submit" name="searchbtn" class="removebtnstyle"><img src="/views/imgs/icons/search.png"
-                        class="icon"></button>
+                <button type="submit" name="searchbtn" class="removebtnstyle"><img src="/views/imgs/icons/search.png" class="icon"></button>
             </form>
         </div>
         <!-- <form method="POST" name="multiactions" action="">
@@ -51,18 +50,17 @@ $countusers = $db->read('users');
     </div>
     <div class="container mt-2">
         <div class="">
-            <a class="btn btn-light mb-1 border border-1 border-success mx-1" href="usernew">Add User
+            <a class="btn btn-light mb-1 border border-2 border-success mx-1" href="usernew">Add User
                 <img src="/views/imgs/icons/adduser.png" width="30" height="30">
             </a>
         </div>
         <table class="table table-striped rounded rounded-5 overflow-hidden text-center">
-            <thead class="bg-light text-success border border-3 border-success">
+            <thead class="bg-success text-light">
                 <tr>
                     <th>#</th>
                     <th></th>
                     <th>User/Badge</th>
                     <th>Full Name</th>
-                    <th>Employment Type</th>
                     <th>Approved</th>
                     <th>Joining Date</th>
                     <th>Vacation Balance</th>
@@ -89,45 +87,37 @@ $countusers = $db->read('users');
             $users = $db->read('users', $where, $params, $limit, null, 'id');
             foreach ($users as $user) {
             ?>
-            <!-- disabe if not author or admin -->
-            <?php $disable =  authBtns($_SESSION['uid'], $user['id'], $_SESSION['role']); ?>
-            <?php $adminDisable =  isadmin($_SESSION['role']); ?>
+                <!-- disabe if not author or admin -->
+                <?php $disable =  authBtns($_SESSION['uid'], $user['id'], $_SESSION['role']); ?>
+                <?php $adminDisable =  isadmin($_SESSION['role']); ?>
 
-            <?php if ($user['approved'] == "0") {
+                <?php if ($user['approved'] == "0") {
                     $class = "linethru";
                 } else {
                     $class = "";
                 } ?>
-            <tbody class="">
+
                 <tr class="<?= $class ?>">
                     <td><?= $user['id']; ?></td>
                     <td>
-                        <img src="<?= ($user['image'] ?? "views/imgs/guest.png") ?>" alt="<?= $user['name']; ?>"
-                            class="userimg scale2">
+                        <img src="<?= ($user['image'] ?? "views/imgs/guest.png") ?>" alt="<?= $user['name']; ?>" class="userimg scale2">
                     </td>
                     <td><?= $user['user']; ?></td>
                     <td><a href="userview?id=<?= $user['id']; ?>"><?= $user['name']; ?></a></td>
-                    <td><?= $user['emptype']; ?></td>
                     <td><?= $user['approved'] == '1' ? 'Yes' : 'No'; ?></td>
                     <td><?= $user['joiningdate']; ?></td>
                     <td><?= $user['vacbalance']; ?></td>
 
                     <td>
 
-                        <a title="Edit" href="userupdate?id=<?= $user['id'] ?>" class="<?= $disable ?>"><img
-                                class="icon" src="views/imgs/icons/edit.png"></a>
-                        <a title="Delete" onclick="return confirm('Are you sure you want to Delete?')"
-                            href="userdelete?id=<?= $user['id'] ?>" class="<?= $adminDisable ?>"><img class="icon"
-                                src="views/imgs/icons/delete.png"></a>
+                        <a title="Edit" href="userupdate?id=<?= $user['id'] ?>" class="<?= $disable ?>"><img class="icon" src="views/imgs/icons/edit.png"></a>
+                        <a title="Delete" onclick="return confirm('Are you sure you want to Delete?')" href="userdelete?id=<?= $user['id'] ?>" class="<?= $adminDisable ?>"><img class="icon" src="views/imgs/icons/delete.png"></a>
                         <?php if ($user['approved'] == 1) { ?>
-                        <a title="Block"
-                            onclick="return confirm('Are you sure you want to Block <?= $user['name'] ?> ?')"
-                            href="userblock?id=<?= $user['id'] ?>" class="<?= $adminDisable ?>"><img class="icon"
-                                src="views/imgs/icons/banuser.png"></a>
+                            <a title="Block" onclick="return confirm('Are you sure you want to Block <?= $user['name'] ?> ?')" href="userblock?id=<?= $user['id'] ?>" class="<?= $adminDisable ?>"><img class="icon" src="views/imgs/icons/banuser.png"></a>
                         <?php } ?>
                     </td>
                 </tr>
-            </tbody>
+
             <?php } ?>
         </table>
         <!-- </form> -->
